moved helper logic to another folder since they don't need to be under the same rules as the source logic
